third_party/blink/web_tests/platform/win/fast/forms/select-popup/popup-menu-open-partially-visible-expected.txt

PASS internals.pagePopupWindow.isWindowHidden() is false
PASS successfullyParsed is true

TEST COMPLETE

o1
o2
